refactor(db): extract settings CRUD + library upserts to db/src/settings.rs
