+++ db/src/lib.rs
@@ -14,6 +14,7 @@ pub mod library_layout;
 pub mod pool;
 pub mod queries;
 pub mod scanner;
+pub mod settings;
 mod taxonomy;
 pub mod thumbs;
 pub mod worker;
@@ -25,4 +26,5 @@ pub use app_state::*;
 pub use helpers::{build_fts_match, sanitize_fts_query};
 pub use pool::*;
 pub use queries::*;
+pub use settings::*;
 pub use thumbs::{thumb_path_for, thumbs_dir, ThumbError, ThumbSize};
